viré l'exeption ds le driver memcache qd il arrive pas a se connecter au serveur

+++ lib/core_cacher_memcache.php
@@ -62,16 +62,17 @@ class core_cacher_memcache extends core_cacher_lib {
 			);
 		}
 
-		if(!$conn) {
-			throw new wf_exception(
-				$this,
-				WF_EXC_PRIVATE,
-				'Cannot connect to memcache server ('.$this->host.':'.$this->port.').'
-			);
-		}
+// 		if(!$conn) {
+// 			throw new wf_exception(
+// 				$this,
+// 				WF_EXC_PRIVATE,
+// 				'Cannot connect to memcache server ('.$this->host.':'.$this->port.').'
+// 			);
+// 		}
 	}
 
 	public function __destruct() {
+		if(!$conn) return;
 		/* close non-persistant connection */
 		if(!$this->persistant) {
 			$this->memcache->close();
@@ -79,23 +80,28 @@ class core_cacher_memcache extends core_cacher_lib {
 	}
 	
 	public function store($var, $val, $timeout) {
+		if(!$conn) return;
 		return($this->memcache->set($var, serialize($val), $this->compression, $timeout));
 	}
 	
 	public function get($var) {
+		if(!$conn) return;
 		/* auto deserialization */
  		return($this->memcache->get($var));
 	}
 	
 	public function delete($var) {
+		if(!$conn) return;
 		return($this->memcache->delete($var));
 	}
 
 	public function clear() {
+		if(!$conn) return;
 		return($this->memcache->flush());
 	}
 
 	public function get_banner() {
+		if(!$conn) return('memcache server unreachable on '.$this->host.':'.$this->port);
 		$status = $this->memcache->getServerStatus($this->host);
 		return(
 			'memcache '.$this->memcache->getVersion().' '.
